Commit Kafka offsets in the same thread - Kafka consumer cannot be shared, we can't use an actor here
